GitButler Integration Commit
This is an integration commit for the virtual branches that GitButler is tracking.

Due to GitButler managing multiple virtual branches, you cannot switch back and
forth between git branches and virtual branches easily. 

If you switch to another branch, GitButler will need to be reinitialized.
If you commit on this branch, GitButler will throw it away.

Here are the branches that are currently applied:
 - OSError: [Errno 28] No space left on device
 (refs/gitbutler/OSError-Errno-28-No-space-left-on-device-)
   branch head: a03c14d2fa50970ee6d4d05391a7dbe3e66bce3f
 - Virtual branch (refs/gitbutler/Virtual-branch)

Your previous branch was: refs/heads/main

The sha for that commit was: 37563f60fca5475b046eb33968a652abade38ecf

For more information about what we're doing here, check out our docs:
https://docs.gitbutler.com/features/virtual-branches/integration-branch